Added several useful tools for creating and visualizing beam lines
Added setup_beam_scenario() in xics_rt_tools.py which creates an ideal scenario with the source, graphite, crystal, and detector all lined up according to their bragg angles. Several adjustable distances.

Added a MatPlotLib Plot3D function that displays the location and orientation of the source, graphite, crystal, and detector in a 3D plot.

Added the option to also display the rays as vectors on the plot.