Modest Bugfixing
Made a wide variety of small aterations to bring my code up to date with Novimir's branch. Ray tracer now works properly, but the visualizer is still NYI. Multi-run switching system cooperates well with new misalignment script, so I can now perform many misalignment tests in quick succession.

Created a new folder called "results". This folder will not be included in the commit because it holds all of the .tif files. As luck would have it, the folder is already accounted for in the .gitignore file.