Minor change: plasma temperature data has wrong units
From now on, we should require all temperature data be in electronvolts. Our code runs in eV, but the data that Zhifeng gave us uses keV.

I could've fixed this problem by adding a factor of 1000 to the temperature in the plasma script, but that would've caused many more issues later on, so I decided to avoid taking on the technical debt.