Meshgrid Bugfixes
Fixed a bug where the meshgrid cared about what order you entered the mesh_points in, i.e. entering the points in clockwise order would result in a different triangle normal than entering the points in counterclockwise order. The meshgrid is more robust due to the fix, as now it can reflect x-rays from both sides.

Fixed a bug affecting the meshgrid's intersect check. The intersect check relied on performing a logical operation on floating point numbers, which have float imprecison. This resulted in rays that would normally pass the intersect check being rejected because the sum of three numbers didn't quite add up to 1. This problem was resolved with some rounding.