Changed XICSRT Output
The configuration structure of of the output dictionary doesn't make much sense, so I made it so that the ray tracer outputs its an entire ray history, much like it used to. This has the added benefit of cooperating better with the ipyvolume code and makes the regular visualizer less likely to glitch out and display no rays.

The output dictionary is rather large, however, as it contains copies of the the rays dictionaries at each step in the ray tracing process. It's up to the other python scripts to process the output into something more mangeable, and all of them already do that.