super() update undone
It turns out that me upgrading Spyder to version 4.0.0 resulted in the creation of a new environment on my computer that uses Python 2.7.0, which is why the code would not work until I changed all of the super() calls. After figuring everything out and upgrading the environment to Python 3.7.0, I can now safely revert the changes to the code's super() calls.

Another small change: output will now always have around 5000 entries, regardless of the number of runs. That's because I've normalized the random ray cutter to divide by the number of runs.