Added xics_rt_meshes.py
This script generates the sinusoidal spiral geometry that Manfred asked us about. It's still in very early stages of development and still needs a lot of polish to get it up-to-standard with the rest of the code, but it does generate a the mesh points correctly. It currently needs a new script to solve for the mesh faces, the one it has right now does not work well.

Other than that, I made a few minor changes to some other parts of the script and regenerated the input json file.