Fixed RealPlasma
Used a try/except block to make the RealPlasma properly manage bundles that generate outside the ITER plasma. Also added a try/except block to the raytracer to make sure it doesn't fail when there are no rays generated.

The ray statistics printout at the end of the ray tracing calculation divides by the number of rays generated. If no rays are generated, then this results in a DivideByZeroError. The code now properly handles this exception.