Partial validation check
Got started on verifying that the model simulates rays correctly. Effort was made in ensuring that the graphite mesh reflects rays exactly the same way as the graphite rectangle. This involved making a few changes to the way the optics handle random numbers: the state of the random number generator is now the same whether use_mishgrid is on or off. Right now, the main problem concerning meshes involves the mesh intersect check overcounting a few rays, resulting in a propogating error.

Also made a few other alterations to facilitate the validation check.